+++ lib/logging/logging.js
@@ -18,14 +18,16 @@ const
 winston = require("winston"),
 FileLogger = require('./winston-file'),
 HekaLogger = require('./winston-heka'),
-StatsdLogger = require('./winston-statsd');
+StatsdLogger = require('./winston-statsd'),
+MetricsLogger = require('./winston-metrics');
 
 
 exports.logger = new (winston.Logger)({
   transports: [
     new FileLogger({}),
     new HekaLogger({}),
-    new StatsdLogger({})
+    new StatsdLogger({}),
+    new MetricsLogger({})
   ]
 });
