+++ lib/logging/metrics.js
@@ -18,61 +18,12 @@
  */
 
 const
-winston = require("winston"),
-configuration = require("../configuration"),
-path = require('path'),
-fs = require('fs'),
+logger = require("./logging").logger,
 urlparse = require('urlparse');
-
-// existsSync moved from path in 0.6.x to fs in 0.8.x
-if (typeof fs.existsSync === 'function') {
-  var existsSync = fs.existsSync;
-} else {
-  var existsSync = path.existsSync;
-}
-
-// go through the configuration and determine log location
-// for now we only log to one place
-// FIXME: separate logs depending on purpose?
-
-var log_path = path.join(configuration.get('var_path'), 'log');
-var LOGGER;
-
-// simple inline function for creation of dirs
-function mkdir_p(p) {
-  if (!existsSync(p)) {
-    mkdir_p(path.dirname(p));
-    fs.mkdirSync(p, "0755");
-  }
-}
-
-function setupLogger() {
-  // don't create the logger if it already exists
-  if (LOGGER) return;
-
-  if (!log_path)
-    return console.log("no log path! Not logging!");
-  else
-    mkdir_p(log_path);
-
-  var filename = path.join(log_path, configuration.get('process_type') + "-metrics.json");
-  if (process.env.METRICS_LOG_FILE) {
-    filename = process.env.METRICS_LOG_FILE;
-  }
-
-  LOGGER = new (winston.Logger)({
-      transports: [new (winston.transports.File)({filename: filename})],
-      timestamp: function () { return new Date().toISOString(); },
-    });
-}
 
 // entry is an object that will get JSON'ified
 exports.report = function(type, entry) {
-  // setup the logger if need be
-  setupLogger();
-
-  // allow convenient reporting of atoms by converting
-  // atoms into objects
+  // allow convenient reporting of atoms by converting atoms into objects
   if (entry === null || typeof entry !== 'object') entry = { msg: entry };
   if (entry.type) throw "reported metrics may not have a `type` property, that's reserved";
   entry.type = type;
@@ -81,8 +32,7 @@ exports.report = function(type, entry) {
   if (entry.at) throw "reported metrics may not have an `at` property, that's reserved";
   entry.at = new Date().toUTCString();
 
-  // if no logger, go to console (FIXME: do we really want to log to console?)
-  LOGGER.info(JSON.stringify(entry));
+  logger.info('metrics.report', JSON.stringify(entry));
 };
 
 // utility function to log a bunch of stuff at user entry point
